Creating example scripts
some example scripts
