--- Stylo Gestures/Assets/StyloGestures/Resources/Prefabs/DegreeAngleSwipeDetection.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a2872882649ba4a2d881b1a8cb0aad27
-timeCreated: 1452732151
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
