--- Stylo Gestures/Assets/StyloGestures/Resources/Prefabs/DoubleTapDetection.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 885bf762933f04d74a1bac2ffdb0921a
-timeCreated: 1453147376
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
