--- Stylo Gestures/Assets/StyloGestures/Resources/Prefabs/NormalizedVectorSwipeDetection.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 4afcffaae9e634f5eb6e303eed0a63b9
-timeCreated: 1452732334
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
