--- Stylo Gestures/Assets/StyloGestures/Resources/Prefabs/RadianAngleSwipeDetection.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 2c2b49b08c3714ba4b26b40f8cf1ce8c
-timeCreated: 1452732210
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
