--- Stylo Gestures/Assets/StyloGestures/Resources/Prefabs/RawVectorSwipeDetection.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dd4aa32b384a543f8aa3a157c107f526
-timeCreated: 1452732296
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
