--- Stylo Gestures/Assets/StyloGestures/Resources/Prefabs/SimpleSwipeDetection.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: ec8d1f1938ca74ee4bcf9210aff4b1c3
-timeCreated: 1452732254
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
