--- Stylo Gestures/Assets/StyloGestures/Resources/Prefabs/SingleTapDetection.prefab.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a34fa47c035f64f81bfe5524e1180d98
-timeCreated: 1453147054
-licenseType: Free
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
