+++ Stylo Gestures/Assets/StyloGestures/Scripts/Common.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 0442ba2e7e3274c768ac98970e283cee
+guid: e0001dd5a82d74f3786988e60e606013
 folderAsset: yes
-timeCreated: 1452570272
+timeCreated: 1453325837
 licenseType: Free
 DefaultImporter:
   userData:
