+++ Stylo Gestures/Assets/StyloGestures/Scripts/Common/GestureEnumeration.cs.meta
