+++ Stylo Gestures/Assets/StyloGestures/Scripts/Common/GestureParameters.cs.meta
