+++ Stylo Gestures/Assets/StyloGestures/Scripts/Drag/DragDetectionExample.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace StyloGestures{
+
+	public class DragDetectionExample : DragGesture {
+
+		public override void OnDraggingDetected (Vector2 screenPosition, Vector2 direction)
+		{
+			Debug.Log("Actual touch position : " + screenPosition + " direction: " + direction);
+		}
+	}
+}
