+++ Stylo Gestures/Assets/StyloGestures/Scripts/Drag/DragDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: e19bfb1f066f9474d84f1e825581a45f
-timeCreated: 1453138519
+guid: f77697de4e6054988a70a973e52d486d
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
