--- Stylo Gestures/Assets/StyloGestures/Scripts/Drag/LongPressDetection.cs
@@ -1,12 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-using StyloGestures;
-
-public class LongPressDetection : DragGesture {
-
-	public override void OnLongPressDetected (Vector2 screenPosition)
-	{
-		Debug.Log("Actual touch position: " + screenPosition);
-	}
-
-}
