+++ Stylo Gestures/Assets/StyloGestures/Scripts/Drag/LongPressDetectionExample.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace StyloGestures{
+
+	public class LongPressDetectionExample : DragGesture {
+
+		public override void OnLongPressDetected (Vector2 screenPosition)
+		{
+			Debug.Log("Actual touch position: " + screenPosition);
+		}
+
+	}
+}
