+++ Stylo Gestures/Assets/StyloGestures/Scripts/Drag/LongPressDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b465521e47cb0452084dbb58872d046c
-timeCreated: 1452731264
+guid: c8dce2b523acc4adf95f99b31b53a5e8
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
