+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/DegreeSwipeDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: e2967f71eba4d4d4ab12e31c80cd8348
-timeCreated: 1452731449
+guid: 3d144ba96973e40f68a380fa6f047f27
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
