+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/NormalizedSwipeDetectionExample.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace StyloGestures{
+
+	public class NormalizedSwipeDetectionExample : SwipeGesture {
+
+		public override void OnSwipeNormalizedDetected (Vector2 direction)
+		{
+			Debug.Log("Normalized direction " + direction);
+		}
+
+	}
+}
