+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/NormalizedSwipeDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4d6ccefc9e3d54ff98bc9ea27081f653
-timeCreated: 1453147289
+guid: f4e7e300fae614ed592da5725238eff4
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
