+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/RadianSwipeDetectionExample.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace StyloGestures{
+
+	public class RadianSwipeDetectionExample : SwipeGesture {
+
+		public override void OnSwipeRadianDetected (float angle, float radius)
+		{
+			Debug.Log("Radian Angle : " + angle + " radius: " + radius);
+		}
+
+	}
+}
