+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/RadianSwipeDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: aa360f4c6344d4fa4ae4fd5197463f5a
-timeCreated: 1452731382
+guid: e44a31addb28d4d99963c9350759682d
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
