+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/RawSwipeDetectionExample.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace StyloGestures{
+
+	public class RawSwipeDetectionExample : SwipeGesture {
+
+		public override void OnSwipeRawDetected (Vector2 direction)
+		{
+			Debug.Log("Raw direction " + direction);	
+		}
+		
+	}
+}
