+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/RawSwipeDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 4fb577c857f2e49008c778005b2a15c5
-timeCreated: 1452733300
+guid: e4f563c95ee444bd587409c75ac60909
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
