+++ Stylo Gestures/Assets/StyloGestures/Scripts/Swipe/SimpleSwipeDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: effaa07794b2442878629d24bc558c67
-timeCreated: 1453325149
+guid: 1e7f19460d4f64c6eb0ca0d6a40e28d6
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
