+++ Stylo Gestures/Assets/StyloGestures/Scripts/Tap/DoubleTapDetectionExample.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace StyloGestures{
+
+	public class DoubleTapDetectionExample : TapGesture {
+
+		public override void DoubleTapDetected ()
+		{
+			Debug.Log("Double Tap!");
+		}
+	}
+}
