+++ Stylo Gestures/Assets/StyloGestures/Scripts/Tap/DoubleTapDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 6529ad15b57ae454ea04bb74d8f3c5f9
-timeCreated: 1452733277
+guid: 31b9dd9a3ec714da29cd2db7c4617f68
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
