--- Stylo Gestures/Assets/StyloGestures/Scripts/Tap/SingleTapDetection.cs
@@ -1,13 +1,0 @@
-﻿using UnityEngine;
-using System.Collections;
-using StyloGestures;
-
-[AddComponentMenu("StyloGestures/Tap Gesture/Single Tap")]
-public class SingleTapDetection : TapGesture {
-
-	public override void SingleTapDetected ()
-	{ 
-		Debug.Log("Simple tap!");
-	}
-
-}
