+++ Stylo Gestures/Assets/StyloGestures/Scripts/Tap/SingleTapDetectionExample.cs
@@ -1,0 +1,14 @@
+﻿using UnityEngine;
+using System.Collections;
+
+namespace StyloGestures{
+
+	public class SingleTapDetectionExample : TapGesture {
+
+		public override void SingleTapDetected ()
+		{ 
+			Debug.Log("Simple tap!");
+		}
+
+	}
+}
