+++ Stylo Gestures/Assets/StyloGestures/Scripts/Tap/SingleTapDetectionExample.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 428bbec48be8d4e7d96702c4dd7e2686
-timeCreated: 1453323770
+guid: 976d0911c90984cf7a79a39bf6583f0a
+timeCreated: 1453325796
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
